Add files via upload
The following images are the results of running a 30 day forecast on V1 of the GenerationPredictor machine learning model with a 0.2 test size, and 0.15 validation size.

All results are trained and tested on four distinct machine learning models with the best result being shown, as of V1 the best machine learning model for this setup is the Elastic training model (as seen in the Demand Forecast Results for All 4 Models image) and thus is the one included in the above images. As the code becomes more optimized and features are added, the amount of models that are used might be adjusted with a potential singularization towards a single model.

For future updates, we are currently aiming to produce a standalone application that will be able to properly display the results of the forecaster in more user friendly formats.